+++ tests/PhpSpreadsheetTests/Calculation/Functions/Financial/PmtTest.php
@@ -3,16 +3,10 @@
 namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Financial;
 
 use PhpOffice\PhpSpreadsheet\Calculation\Financial;
-use PhpOffice\PhpSpreadsheet\Calculation\Functions;
 use PHPUnit\Framework\TestCase;
 
 class PmtTest extends TestCase
 {
-    protected function setUp(): void
-    {
-        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
-    }
-
     /**
      * @dataProvider providerPMT
      *
@@ -23,7 +17,13 @@ class PmtTest extends TestCase
         $interestRate = array_shift($args);
         $numberOfPeriods = array_shift($args);
         $presentValue = array_shift($args);
-        $result = Financial::PMT($interestRate, $numberOfPeriods, $presentValue, ...$args);
+        if (count($args) === 0) {
+            $result = Financial::PMT($interestRate, $numberOfPeriods, $presentValue);
+        } elseif (count($args) === 1) {
+            $result = Financial::PMT($interestRate, $numberOfPeriods, $presentValue, $args[0]);
+        } else {
+            $result = Financial::PMT($interestRate, $numberOfPeriods, $presentValue, $args[0], $args[1]);
+        }
         self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
     }